second HM (reworked)

+++ src/HomeWork2/Average.java
@@ -3,13 +3,13 @@ package HomeWork2;
 public class Average {
     public static void main(String[] args) {
 
-        int [] array = {1, 7, 10, -10};
-        int sum = 0;
+        int[] array = {1, 7, 10, -10};
+        double sum = 0;
 
         for (int i : array) {
-            sum+= i;
+            sum += i;
         }
-        int avg = sum / array.length;
+        double avg = sum / array.length;
         System.out.println(avg);
     }
 }
